+++ cmd/parser/main.go
@@ -5,6 +5,8 @@ import (
 	"os"
 	"os/signal"
 	"syscall"
+
+	"github.com/spf13/cobra"
 )
 
 var gitCommit string
@@ -26,4 +28,18 @@ func setupCloseHandler() {
 		fmt.Printf("\r- Signal '%v' received from Terminal. Exiting...\n ", sig)
 		os.Exit(0)
 	}()
+}
+
+var (
+	// Used for flags.
+	rootCmd = &cobra.Command{
+		Use:     "hap",
+		Short:   "A tool for parsing and validation of HAP rules",
+		Version: gitCommit,
+		Long:    ``,
+	}
+)
+
+func Execute() error {
+	return rootCmd.Execute()
 }
